Prevent drag and drop for hidden widgets
We’re filtering out hidden target elements here instead of when collecting available drop targets, because a widget visibility can be changed while the drag and drop is in progress. If we only checked the visibility once when creating the drag and drop state, we would not be able to handle that use case.